michalskiba_dev/requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
click==8.1.3
coverage==7.2.1
Django==4.1.5
django-factory-boy==1.0.0
django-stubs==1.16.0
django-stubs-ext==0.8.0
factory-boy==3.2.1
Faker==18.3.0
gunicorn==20.1.0
iniconfig==2.0.0
isort==5.12.0
mypy==1.1.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-slugify==8.0.1
ruff==0.0.254
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
types-python-slugify==8.0.0.2
types-pytz==2022.7.1.2
types-PyYAML==6.0.12.8
typing_extensions==4.5.0
